Refactor `deepseg_` files into shared/common folder
