+++ spinalcordtoolbox/deepseg_/lesion.py
@@ -11,7 +11,7 @@ import nibabel as nib
 
 from spinalcordtoolbox.image import Image, add_suffix, zeros_like, empty_like
 from spinalcordtoolbox.deepseg.models import onnx_inference
-from spinalcordtoolbox.deepseg_sc.core import find_centerline, crop_image_around_centerline, uncrop_image, _normalize_data
+from spinalcordtoolbox.deepseg_.sc import find_centerline, crop_image_around_centerline, uncrop_image, _normalize_data
 from spinalcordtoolbox import resampling
 from spinalcordtoolbox.utils import sct_dir_local_path, TempFolder
