+++ spinalcordtoolbox/scripts/sct_deepseg_lesion.py
@@ -143,7 +143,7 @@ def main(argv=None):
 
     # Segment image
     from spinalcordtoolbox.image import Image
-    from spinalcordtoolbox.deepseg_lesion.core import deep_segmentation_MSlesion
+    from spinalcordtoolbox.deepseg_.lesion import deep_segmentation_MSlesion
     im_image = Image(fname_image)
     im_seg, im_labels_viewer, im_ctr = deep_segmentation_MSlesion(im_image, contrast_type, ctr_algo=ctr_algo, ctr_file=manual_centerline_fname,
                                         brain_bool=brain_bool, remove_temp_files=remove_temp_files, verbose=verbose)
